requirements.txt
fastapi
uvicorn
python-docx
requests
pydantic
google-genai
python-multipart
fpdf2
pytest
httpx
apscheduler
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pyjwt
email-validator
stripe
